20250518_bright_data_mcp_final_2_prod_1/bright_data_example.py
import os
from dotenv import load_dotenv
import json
import pandas as pd

# Load environment variables
load_dotenv()

# This file demonstrates how to use Bright Data MCP for web scraping
# In a real implementation, you would use the run_mcp tool to interact with Bright Data

def demonstrate_bright_data_search():
    """Demonstrate how to use Bright Data MCP for searching.
    
    This function shows the structure of how you would use the Bright Data MCP
    to search for robot vacuum cleaners on Amazon Italy.
    
    Returns:
        dict: Example of what the search results would look like
    """
    print("Demonstrating Bright Data MCP search...")
    
    # In a real implementation, you would use the run_mcp tool like this:
    # search_results = run_mcp({
    #     "server_name": "mcp.config.usrlocalmcp.Bright Data",
    #     "tool_name": "search_engine",
    #     "args": {
    #         "query": "robot aspirapolvere amazon.it",
    #         "engine": "google"
    #     }
    # })
    
    # Example of what the search results might look like
    example_search_results = {
        "results": [
            {
                "title": "Robot Aspirapolvere - Amazon.it",
                "url": "https://www.amazon.it/s?k=robot+aspirapolvere",
                "description": "Risultati per robot aspirapolvere su Amazon.it"
            },
            {
                "title": "Robot Aspirapolvere e Lavapavimenti - Amazon.it",
                "url": "https://www.amazon.it/s?k=robot+aspirapolvere+lavapavimenti",
                "description": "Scopri i migliori robot aspirapolvere e lavapavimenti su Amazon.it"
            }
        ]
    }
    
    return example_search_results

def demonstrate_bright_data_scraping():
    """Demonstrate how to use Bright Data MCP for scraping product pages.
    
    This function shows the structure of how you would use the Bright Data MCP
    to scrape product details from Amazon Italy.
    
    Returns:
        dict: Example of what the scraped product data would look like
    """
    print("Demonstrating Bright Data MCP scraping...")
    
    # In a real implementation, you would use the run_mcp tool like this:
    # product_data = run_mcp({
    #     "server_name": "mcp.config.usrlocalmcp.Bright Data",
    #     "tool_name": "scrape_as_markdown",
    #     "args": {
    #         "url": "https://www.amazon.it/dp/B0BVMKJH7L"
    #     }
    # })
    
    # Example of what the scraped product data might look like
    example_product_data = {
        "content": "# iRobot Roomba Combo Essential Robot Aspirapolvere e Lavapavimenti 2-in-1\n\n**Prezzo:** €249,00\n\n**Valutazione:** 4,3 su 5 stelle (1.245 recensioni)\n\n**Caratteristiche principali:**\n- Sistema di pulizia a 2 fasi\n- Navigazione intelligente\n- Compatibile con assistenti vocali\n- Autonomia fino a 110 minuti\n- Tecnologia Dirt Detect\n\n**Disponibilità:** In stock\n\n**Spedizione:** Prime\n"
    }
    
    return example_product_data

def parse_markdown_to_structured_data(markdown_content):
    """Parse markdown content to structured data.
    
    This function demonstrates how you would parse the markdown content
    returned by Bright Data MCP into structured data for analysis.
    
    Args:
        markdown_content (str): Markdown content from Bright Data MCP
        
    Returns:
        dict: Structured product data
    """
    print("Parsing markdown content to structured data...")
    
    # This is a simplified example of parsing
    # In a real implementation, you would use regex or a markdown parser
    
    lines = markdown_content.split('\n')
    
    # Extract title (first line after removing # )
    title = lines[0].replace('# ', '') if lines[0].startswith('# ') else ''
    
    # Extract price
    price = ''
    for line in lines:
        if '**Prezzo:**' in line:
            price = line.split('**Prezzo:**')[1].strip()
            break
    
    # Extract rating
    rating = ''
    for line in lines:
        if '**Valutazione:**' in line:
            rating_part = line.split('**Valutazione:**')[1].strip()
            rating = rating_part.split(' ')[0]
            break
    
    # Extract reviews count
    reviews = ''
    for line in lines:
        if 'recensioni' in line:
            reviews_part = line.split('(')[1].split(')')[0] if '(' in line and ')' in line else ''
            reviews = reviews_part
            break
    
    # Extract features
    features = []
    feature_section = False
    for line in lines:
        if '**Caratteristiche principali:**' in line:
            feature_section = True
            continue
        if feature_section and line.startswith('- '):
            features.append(line.replace('- ', ''))
        elif feature_section and line.startswith('**'):
            feature_section = False
    
    # Create structured data
    structured_data = {
        "title": title,
        "price": price,
        "rating": rating,
        "reviews": reviews,
        "features": features
    }
    
    return structured_data

def main():
    """Main function to demonstrate Bright Data MCP usage."""
    print("Starting Bright Data MCP demonstration")
    
    # Step 1: Demonstrate search
    search_results = demonstrate_bright_data_search()
    print("\nExample search results:")
    print(json.dumps(search_results, indent=2))
    
    # Step 2: Demonstrate scraping
    product_data = demonstrate_bright_data_scraping()
    print("\nExample scraped product data (markdown):")
    print(product_data["content"])
    
    # Step 3: Demonstrate parsing
    structured_data = parse_markdown_to_structured_data(product_data["content"])
    print("\nParsed structured data:")
    print(json.dumps(structured_data, indent=2))
    
    print("\nIn a real implementation, you would:")
    print("1. Use the search results to find product URLs")
    print("2. Scrape each product URL to get detailed information")
    print("3. Parse the scraped data into structured format")
    print("4. Analyze the data and create visualizations")
    
    print("\nDemonstration complete!")

if __name__ == "__main__":
    main()